src/test/java/org/example/clientinfo/exception/GlobalExceptionHandlerTest.java
package org.example.clientinfo.exception;

import org.junit.Assert;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.mockito.InjectMocks;
import org.mockito.Mockito;
import org.mockito.junit.MockitoJUnitRunner;
import org.springframework.core.MethodParameter;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.MethodArgumentNotValidException;
import org.springframework.web.context.request.WebRequest;

@RunWith(MockitoJUnitRunner.class)
public class GlobalExceptionHandlerTest {

    @InjectMocks
    GlobalExceptionHandler globalExceptionHandler;

    @Test
    public void testExceptionForApiException(){
        ResponseEntity<?> response = globalExceptionHandler.customValidationErrorHandling(new ApiException("Test Error"));
        Assert.assertNotNull(response);
        Assert.assertEquals(HttpStatus.BAD_REQUEST, response.getStatusCode());

    }

    @Test
    public void testExceptionForResourceNotFoundException(){
        ResponseEntity<?> response = globalExceptionHandler.resourceNotFoundHandling(new ResourceNotFoundException("Test Error"));
        Assert.assertNotNull(response);
        Assert.assertEquals(HttpStatus.NOT_FOUND, response.getStatusCode());
    }

}
